Module fsort::criterion

This module contains different criteria for sorting files.

Structs

AccessDate
A criterion which sorts files by the dates of the last access to them.

CreationDate
A criterion which sorts files by their dates of creation.

FileName
A criterion which sorts files by their names.

FileSize
A criterion which sorts files by their size.

ModifyDate
A criterion which sorts files by their last dates of modify.

Enums

Error
An error occurring during the creation of a criterion.

Traits

Criterion
A trait for objects which are valid sorting criteria for files.