Module fsort::file_collection

This modules contains the sorted FileCollections and their iterators.

Structs

DynamicCollection
A sorted collection of files with changeable criterion.

FileIter
An iterator over the files in a FileCollection according to the requested order.

PathIter
An iterator over the sorted paths of a FileCollection.

StaticCollection
A sorted collection of files with fixed criterion.

Enums

ErrorBehavior
This enum describes the behavior of a sorting algorithm in case of errors while adding files.

Result
The enum describes the result of a file adding operation.

Traits

FileCollection
A trait for objects which are sorted collections of files.