Struct fsort::file_collection::StaticCollection

pub struct StaticCollection<C: Criterion> {
    // some fields omitted
}

A sorted collection of files with fixed criterion.

Methods

impl<C: Criterion> StaticCollection<C>

fn new() -> StaticCollection<C>

Creates a new sorted collection with a specific criterion.

Examples

use fsort::criterion::FileName;
use fsort::file_collection::StaticCollection;

let collection = StaticCollection::<FileName>::new();

fn iter(&self) -> FileIter<C>

Returns an iterator over the sorted files in the collection according to the criterion.

Examples

use std::fs::File;
use std::path::PathBuf;
use fsort::criterion::FileName;
use fsort::file_collection::StaticCollection;
use fsort::file_collection::FileCollection;

let mut i1 = std::env::temp_dir();
let mut i2 = std::env::temp_dir();
i1.push("I1.tmp");
i2.push("I2.tmp");
File::create(&i1).unwrap();
File::create(&i2).unwrap();

// Inserts files into collection
let mut collection = StaticCollection::<FileName>::new();
collection.add_file(&i2);
collection.add_file(&i1);

let mut iter = collection.iter();
assert_eq!(i1, iter.next().unwrap().path);
assert_eq!(i2, iter.next().unwrap().path);
assert_eq!(None, iter.next());